Implemented #64

+++ neon/Service/DataGateway.swift
@@ -13,19 +13,7 @@ class DataGateway {
     
     static let shared = DataGateway()
     
-    func checkConnection(completion: @escaping (_ success: Bool) -> ()) {
-        let database = CKContainer.default().privateCloudDatabase
-        let query = CKQuery(recordType: "AgendaRecord", predicate: NSPredicate(value: true))
-        
-        database.perform(query, inZoneWith: nil) { (records, error) in
-            if error != nil {
-                print("ICLOUD ISSUE \(error?.localizedDescription)")
-            }
-            completion(error == nil)
-        }
-    }
-    
-    func fetchAgendaItems(completion: @escaping (_ todaysAgendaItems: [Int: AgendaItem], _ tomorrowsAgendaItems: [Int: AgendaItem]) -> ()) {
+    func fetchAgendaItems(completion: @escaping (_ todaysAgendaItems: [Int: AgendaItem], _ tomorrowsAgendaItems: [Int: AgendaItem], _ success: Bool) -> ()) {
         var todaysAgendaItems = [Int: AgendaItem]()
         var tomorrowsAgendaItems = [Int: AgendaItem]()
         
@@ -51,21 +39,25 @@ class DataGateway {
         let query = CKQuery(recordType: "AgendaRecord", predicate: NSPredicate(value: true))
         
         database.perform(query, inZoneWith: nil) { (records, error) in
-            records?.forEach({ (record) in
-                guard let id = record.value(forKey: "id") as? String else { return }
-                guard let title = record.value(forKey: "title") as? String else { return }
-                guard let hour = record.value(forKey: "hour") as? Int else { return }
-                guard let day = record.value(forKey: "day") as? Date else { return }
+            if error == nil {
+                records?.forEach({ (record) in
+                    guard let id = record.value(forKey: "id") as? String else { return }
+                    guard let title = record.value(forKey: "title") as? String else { return }
+                    guard let hour = record.value(forKey: "hour") as? Int else { return }
+                    guard let day = record.value(forKey: "day") as? Date else { return }
+                    
+                    // Only pull the tasks that are in today and aren't already on device
+                    if Calendar.current.isDateInToday(day) {
+                        todaysAgendaItems[hour] = AgendaItem(with: id, and: title)
+                    } else if Calendar.current.isDateInTomorrow(day) {
+                        tomorrowsAgendaItems[hour] = AgendaItem(with: id, and: title)
+                    }
+                })
                 
-                // Only pull the tasks that are in today and aren't already on device
-                if Calendar.current.isDateInToday(day) {
-                    todaysAgendaItems[hour] = AgendaItem(with: id, and: title)
-                } else if Calendar.current.isDateInTomorrow(day) {
-                    tomorrowsAgendaItems[hour] = AgendaItem(with: id, and: title)
-                }
-            })
-            
-            completion(todaysAgendaItems, tomorrowsAgendaItems)
+                completion(todaysAgendaItems, tomorrowsAgendaItems, true)
+            } else {
+                completion(todaysAgendaItems, tomorrowsAgendaItems, false)
+            }
         }
     }
     
@@ -85,11 +77,14 @@ class DataGateway {
     
     func delete(_ agendaItem: AgendaItem) {
         let database = CKContainer.default().privateCloudDatabase
-        let query = CKQuery(recordType: "AgendaRecord", predicate: NSPredicate(format: "id == %@", agendaItem.id))
+        let query = CKQuery(recordType: "AgendaRecord", predicate: NSPredicate(value: true))
         
         database.perform(query, inZoneWith: nil) { (records, error) in
             records?.forEach({ (record) in
-                database.delete(withRecordID: record.recordID, completionHandler: { (recordID, error) in })
+                guard let id = record.value(forKey: "id") as? String else { return }
+                if id == agendaItem.id {
+                    database.delete(withRecordID: record.recordID, completionHandler: { (recordID, error) in })
+                }
             })
         }
     }
